refactor(core): eliminate three circular import edges (DIP violations) (#1937)
- Promote TokenStats from scylla.e2e.models_internals to scylla.core.token_stats
- Promote progress types from scylla.cli.progress to scylla.core.progress
- Repoint scylla.e2e.progress to scylla.core.progress (breaks cli↔e2e cycle)
- Repoint scylla.adapters.base to scylla.core.token_stats (breaks adapters↔e2e cycle)
- Retarget analysis/loader_internals TokenStats imports to scylla.core.token_stats
- Add back-compat re-export shims in cli/progress.py and e2e/models_internals/token_stats.py
- Retarget test mock patch sites to scylla.core.progress.datetime
- Add import-linter contracts (3 forbidden) in pyproject.toml with pixi lint feature
- Add pixi lint-imports task and CI step in _required.yml
- Add tests/test_import_layering.py pytest wrapper for import-linter contracts

Edge 1 (config↔metrics): not a live cycle; locked by import-linter contracts only.
Edge 2 (cli↔e2e): eliminated — e2e.progress no longer imports from scylla.cli.
Edge 3 (adapters↔e2e): eliminated — adapters.base no longer imports from scylla.e2e.
Note: adapters.claude_code→e2e.rate_limit pre-existing dep narrowed in contract.
