+++ src/scylla/e2e/progress.py
@@ -2,17 +2,29 @@
 
 .. deprecated::
     This module is a backward-compatibility shim.  The canonical
-    location is :mod:`scylla.cli.progress`.  Import from there instead.
+    location is :mod:`scylla.core.progress`.  Import from there instead.
 """
 
-from scylla.cli.progress import (
-    EvalProgress,
-    ProgressDisplay,
-    RunProgress,
-    RunStatus,
-    TierProgress,
-    format_duration,
-    format_progress_bar,
+from scylla.core.progress import (
+    EvalProgress as EvalProgress,
+)
+from scylla.core.progress import (
+    ProgressDisplay as ProgressDisplay,
+)
+from scylla.core.progress import (
+    RunProgress as RunProgress,
+)
+from scylla.core.progress import (
+    RunStatus as RunStatus,
+)
+from scylla.core.progress import (
+    TierProgress as TierProgress,
+)
+from scylla.core.progress import (
+    format_duration as format_duration,
+)
+from scylla.core.progress import (
+    format_progress_bar as format_progress_bar,
 )
 
 __all__ = [
